Make MessageRequest that has required properties
