+++ packages/teams-js/src/internal/interfaces.ts
@@ -39,27 +39,6 @@ export interface ExtendedWindow extends Window {
 }
 
 /**
- * @internal
- * Limited to Microsoft-internal use
- */
-export interface MessageRequest {
-  id?: number;
-  func: string;
-  timestamp?: number;
-  args?: any[];
-}
-
-/**
- * @internal
- * Limited to Microsoft-internal use
- */
-export interface MessageResponse {
-  id: number;
-  args?: any[];
-  isPartialResponse?: boolean; // If the message is partial, then there will be more future responses for the given message ID.
-}
-
-/**
  * @hidden
  * Meant for Message objects that are sent to children without id
  *
