+++ packages/teams-js/test/utils.ts
@@ -1,6 +1,7 @@
 import { defaultSDKVersionForCompatCheck } from '../src/internal/constants';
 import { GlobalVars } from '../src/internal/globalVars';
-import { DOMMessageEvent, ExtendedWindow, MessageResponse } from '../src/internal/interfaces';
+import { DOMMessageEvent, ExtendedWindow } from '../src/internal/interfaces';
+import { MessageResponse } from '../src/internal/messageObjects';
 import { app } from '../src/public/app';
 import { applyRuntimeConfig, IBaseRuntime, setUnitializedRuntime } from '../src/public/runtime';
